Crate tmc_langs_cli

CLI client for TMC

Modules

app

Create clap app

config

Handles the CLI’s configuration files and credentials.

error

output

Output format

Structs

PrintToken

Functions

check_sandbox_err

Goes through the error chain and returns the error output file path if a sandbox test error is found

error_message_special_casing

Goes through the error chain and returns the specialized error message, if any.

into_locale

into_url

into_usize

json_to_toml

main

move_dir

print_output

print_output_with_file

print_warnings

run_app

run_checkstyle_write_results

run_core

run_settings

solve_error_kind

Goes through the error chain and checks for special error types that should be indicated by the Kind.

write_result_to_file_as_json