Enum tmc_langs_cli::app::Command

pub enum Command {

    Checkstyle {
        exercise_path: PathBuf,
        locale: Locale,
        output_path: Option<PathBuf>,
    },
    Clean {
        exercise_path: PathBuf,
    },
    CompressProject {
        exercise_path: PathBuf,
        output_path: PathBuf,
        compression: Compression,
        deterministic: bool,
        naive: bool,
    },
    Tmc(TestMyCode),
    Mooc(Mooc),
    ExtractProject {
        archive_path: PathBuf,
        output_path: PathBuf,
        compression: Compression,
        naive: bool,
    },
    FastAvailablePoints {
        exercise_path: PathBuf,
    },
    FindExercises {
        search_path: PathBuf,
        output_path: Option<PathBuf>,
    },
    GetExercisePackagingConfiguration {
        exercise_path: PathBuf,
        output_path: Option<PathBuf>,
    },
    ListLocalCourseExercises {
        client_name: String,
        course_slug: String,
    },
    PrepareSolution {
        exercise_path: PathBuf,
        output_path: PathBuf,
    },
    PrepareStub {
        exercise_path: PathBuf,
        output_path: PathBuf,
    },
    PrepareSubmission {
        output_format: Compression,
        clone_path: PathBuf,
        output_path: PathBuf,
        stub_archive_path: Option<PathBuf>,
        stub_compression: Compression,
        submission_path: PathBuf,
        submission_compression: Compression,
        extract_submission_naively: bool,
        tmc_param: Vec<String>,
        no_archive_prefix: bool,
    },
    RefreshCourse {
        cache_path: PathBuf,
        cache_root: PathBuf,
        course_name: String,
        git_branch: String,
        source_url: String,
    },
    RunTests {
        checkstyle_output_path: Option<PathBuf>,
        exercise_path: PathBuf,
        locale: Option<Locale>,
        output_path: Option<PathBuf>,
        wait_for_secret: bool,
    },
    Settings(Settings),
    ScanExercise {
        exercise_path: PathBuf,
        output_path: Option<PathBuf>,
    },

}

Variants

Checkstyle

Fields

exercise_path: PathBuf

Path to the directory where the project resides.

locale: Locale

Locale as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

output_path: Option<PathBuf>

If defined, the check results will be written to this path. Overwritten if it already exists.

Checks the code style for the given exercise

Clean

Fields

exercise_path: PathBuf

Path to the directory where the exercise resides.

Cleans the target exercise using the appropriate language plugin

CompressProject

Fields

exercise_path: PathBuf

Path to the directory where the exercise resides.

output_path: PathBuf

Path to the output archive. Overwritten if it already exists.

compression: Compression

Compression algorithm to use.

deterministic: bool

If set, does not include metadata such as timestamps in the archives.

naive: bool

If set, simply compresses the target directory with all of its files.

Compresses the target exercise. Only includes student files using the student file policy of the exercise’s plugin

Tmc(TestMyCode)

Mooc(Mooc)

ExtractProject

Fields

archive_path: PathBuf

Path to the archive.

output_path: PathBuf

Path to the directory where the archive will be extracted.

compression: Compression

Compression algorithm used for the archive.

naive: bool

If set, simply extracts the target directory with all of its files.

Extracts an exercise from an archive. If the output-path is a project root, the plugin’s student file policy will be used to avoid overwriting student files

FastAvailablePoints

Fields

exercise_path: PathBuf

Path to the directory where the projects reside.

Parses @Points notations from an exercise’s exercise files and returns the point names found

FindExercises

Fields

search_path: PathBuf

Path to the directory where to search for exercises.

output_path: Option<PathBuf>

If given, the search results will be written to this path. Overwritten if it already exists.

Finds all exercise root directories inside the exercise-path

GetExercisePackagingConfiguration

Fields

exercise_path: PathBuf

Path to the directory where the exercise resides.

output_path: Option<PathBuf>

If given, the configuration will be written to this path. Overwritten if it already exists.

Returns a configuration which separately lists the student files and exercise files inside the given exercise

ListLocalCourseExercises

Fields

client_name: String

The client name of which the exercises should be listed.

course_slug: String

The course slug the local exercises of which should be listed.

Returns a list of local exercises for the given course

PrepareSolution

Fields

exercise_path: PathBuf

Path to the directory where the exercise resides.

output_path: PathBuf

Path to the directory where the processed files will be written.

Processes the exercise files in exercise-path, removing all code marked as stubs

PrepareStub

Fields

exercise_path: PathBuf

Path to the directory where the exercise resides.

output_path: PathBuf

Path to the directory where the processed files will be written.

Processes the exercise files in exercise-path, removing all code marked as solutions

PrepareSubmission

Fields

output_format: Compression

The output format of the submission archive. Defaults to tar.

clone_path: PathBuf

Path to exercise’s clone path, where the unmodified test files will be copied from. The course and exercise name will be derived from the clone path and used in the resulting submission archive, unless no-archive-prefix is set.

output_path: PathBuf

Path to the resulting archive. Overwritten if it already exists.

stub_archive_path: Option<PathBuf>

If given, the tests will be copied from this stub instead, effectively ignoring hidden tests.

stub_compression: Compression

Compression algorithm used for the stub archive.

submission_path: PathBuf

Path to the submission archive.

submission_compression: Compression

Compression algorithm used for the submission.

extract_submission_naively: bool

If set, the submission is extracted without trying to find a project directory inside it. This can be useful if the submission is minimal and doesn’t contain enough files to detect the project.

tmc_param: Vec<String>

A key-value pair in the form key=value to be written into .tmcparams. If multiple pairs with the same key are given, the values are collected into an array.

no_archive_prefix: bool

If given, the exercise will reside in the root of the archive.

Takes a submission archive and turns it into an archive with reset test files, and tmc-params, ready for further processing. Returns the sandbox image that should be used for the submission.

RefreshCourse

Fields

cache_path: PathBuf

Path to the cached course.

cache_root: PathBuf

The cache root.

course_name: String

The name of the course.

git_branch: String

Version control branch.

source_url: String

Version control URL or path.

Refresh the given course

RunTests

Fields

checkstyle_output_path: Option<PathBuf>

Runs checkstyle if given. Path to the file where the style results will be written.

exercise_path: PathBuf

Path to the directory where the exercise resides.

locale: Option<Locale>

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’. Required if checkstyle-output-path is given.

output_path: Option<PathBuf>

If defined, the test results will be written to this path. Overwritten if it already exists.

wait_for_secret: bool

If defined, the command will wait for a string to be written to stdin, used for signing the output file with jwt.

Run the tests for the exercise using the appropriate language plugin

Settings(Settings)

ScanExercise

Fields

exercise_path: PathBuf

Path to the directory where the project resides.

output_path: Option<PathBuf>

If given, the scan results will be written to this path. Overwritten if it already exists.

Produces a description of an exercise using the appropriate language plugin