Enum tmc_langs_cli::app::Core

pub enum Core {

    CheckExerciseUpdates,
    DownloadModelSolution {
        exercise_id: u32,
        target: PathBuf,
    },
    DownloadOldSubmission {
        submission_id: u32,
        save_old_state: bool,
        exercise_id: u32,
        output_path: PathBuf,
    },
    DownloadOrUpdateCourseExercises {
        download_template: bool,
        exercise_id: Vec<u32>,
    },
    GetCourseData {
        course_id: u32,
    },
    GetCourseDetails {
        course_id: u32,
    },
    GetCourseExercises {
        course_id: u32,
    },
    GetCourseSettings {
        course_id: u32,
    },
    GetCourses {
        organization: String,
    },
    GetExerciseDetails {
        exercise_id: u32,
    },
    GetExerciseSubmissions {
        exercise_id: u32,
    },
    GetExerciseUpdates {
        course_id: u32,
        exercise: Vec<String>,
    },
    GetOrganization {
        organization: String,
    },
    GetOrganizations,
    GetUnreadReviews {
        course_id: u32,
    },
    LoggedIn,
    Login {
        base64: bool,
        email: Option<String>,
        set_access_token: Option<String>,
        stdin: bool,
    },
    Logout,
    MarkReviewAsRead {
        course_id: u32,
        review_id: u32,
    },
    Paste {
        exercise_id: u32,
        locale: Option<Locale>,
        paste_message: Option<String>,
        submission_path: PathBuf,
    },
    RequestCodeReview {
        exercise_id: u32,
        locale: Locale,
        message_for_reviewer: Option<String>,
        submission_path: PathBuf,
    },
    ResetExercise {
        save_old_state: bool,
        exercise_id: u32,
        exercise_path: PathBuf,
    },
    SendFeedback {
        submission_id: Option<u32>,
        feedback_url: Option<String>,
        feedback: Vec<String>,
    },
    Submit {
        dont_block: bool,
        locale: Option<Locale>,
        submission_path: PathBuf,
        exercise_id: u32,
    },
    UpdateExercises,
    WaitForSubmission {
        submission_id: u32,
    },

}

Various commands that communicate with the TMC server

Variants

CheckExerciseUpdates

Checks for updates to any exercises that exist locally.

DownloadModelSolution

Fields

exercise_id: u32

The ID of the exercise.

target: PathBuf

Path to where the model solution will be downloaded.

Downloads an exercise’s model solution

DownloadOldSubmission

Fields

submission_id: u32

The ID of the submission.

save_old_state: bool

If set, the exercise is submitted to the server before resetting it.

exercise_id: u32

The ID of the exercise.

output_path: PathBuf

Path to where the submission should be downloaded.

Downloads an old submission. Resets the exercise at output-path if any, downloading the exercise base from the server. The old submission is then downloaded and extracted on top of the base, using the student file policy to retain student files

DownloadOrUpdateCourseExercises

Fields

download_template: bool

If set, will always download the course template instead of the latest submission, even if one exists.

exercise_id: Vec<u32>

Exercise id of an exercise that should be downloaded. Multiple ids can be given.

Downloads exercises. If downloading an exercise that has been downloaded before, the student file policy will be used to avoid overwriting student files, effectively just updating the exercise files

GetCourseData

Fields

course_id: u32

The ID of the course.

Fetches course data. Combines course details, course exercises and course settings

GetCourseDetails

Fields

course_id: u32

The ID of the course.

Fetches course details

GetCourseExercises

Fields

course_id: u32

The ID of the course.

Lists a course’s exercises

GetCourseSettings

Fields

course_id: u32

The ID of the course.

Fetches course settings

GetCourses

Fields

organization: String

Organization slug (e.g. mooc, hy).

Lists courses

GetExerciseDetails

Fields

exercise_id: u32

The ID of the exercise.

Fetches exercise details

GetExerciseSubmissions

Fields

exercise_id: u32

The ID of the exercise.

Fetches the current user’s old submissions for an exercise

GetExerciseUpdates

Fields

course_id: u32

The ID of the course.

exercise: Vec<String>

An exercise. Takes two values, an exercise id and a checksum. Multiple exercises can be given.

Checks for updates to exercises

GetOrganization

Fields

organization: String

Organization slug (e.g. mooc, hy).

Fetches an organization

GetOrganizations

Fetches a list of all organizations from the TMC server

GetUnreadReviews

Fields

course_id: u32

The ID of the course.

Fetches unread reviews for a course

LoggedIn

Checks if the CLI is authenticated. Prints the access token if so

Login

Fields

base64: bool

If set, the password is expected to be a base64 encoded string. This can be useful if the password contains special characters.

email: Option<String>

The email address of your TMC account. The password will be read through stdin.

set_access_token: Option<String>

The OAUTH2 access token that should be used for authentication.

stdin: bool

If set, the password will be read from stdin instead of TTY like usual. The keyboard input is not hidden in this case, so this should only be used when running the CLI programmatically.

Authenticates with the TMC server and stores the OAuth2 token in config. You can log in either by email and password or an access token

Logout

Logs out and removes the OAuth2 token from config

MarkReviewAsRead

Fields

course_id: u32

The ID of the course.

review_id: u32

The ID of the review.

Marks a review as read

Paste

Fields

exercise_id: u32

The ID of the exercise.

locale: Option<Locale>

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

paste_message: Option<String>

Optional message to attach to the paste.

submission_path: PathBuf

Path to the exercise to be submitted.

Sends an exercise to the TMC pastebin

RequestCodeReview

Fields

exercise_id: u32

The ID of the exercise.

locale: Locale

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

message_for_reviewer: Option<String>

Message for the review.

submission_path: PathBuf

Path to the directory where the submission resides.

Requests code review

ResetExercise

Fields

save_old_state: bool

If set, the exercise is submitted to the server before resetting it.

exercise_id: u32

The ID of the exercise.

exercise_path: PathBuf

Path to the directory where the project resides.

Resets an exercise. Removes the contents of the exercise directory and redownloads it from the server

SendFeedback

Fields

submission_id: Option<u32>

The ID of the submission.

feedback_url: Option<String>

The feedback answer URL.

feedback: Vec<String>

A feedback answer. Takes two values, a feedback answer id and the answer. Multiple feedback arguments can be given.

Sends feedback for an exercise submission

Submit

Fields

dont_block: bool

Set to avoid blocking.

locale: Option<Locale>

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

submission_path: PathBuf

Path to the directory where the exercise resides.

exercise_id: u32

The ID of the exercise.

Submits an exercise. By default blocks until the submission results are returned

UpdateExercises

Updates all local exercises that have been updated on the server

WaitForSubmission

Fields

submission_id: u32

The ID of the submission.

Waits for a submission to finish