tmc_langs_cli::app

Enum TestMyCodeCommand

pub enum TestMyCodeCommand {

    CheckExerciseUpdates,
    DownloadModelSolution {
        exercise_id: u32,
        target: PathBuf,
    },
    DownloadOldSubmission {
        submission_id: u32,
        save_old_state: bool,
        exercise_id: u32,
        output_path: PathBuf,
    },
    DownloadOrUpdateCourseExercises {
        download_template: bool,
        exercise_id: Vec<u32>,
    },
    GetCourseData {
        course_id: u32,
    },
    GetCourseDetails {
        course_id: u32,
    },
    GetCourseExercises {
        course_id: u32,
    },
    GetCourseSettings {
        course_id: u32,
    },
    GetCourses {
        organization: String,
    },
    GetExerciseDetails {
        exercise_id: u32,
    },
    GetExerciseSubmissions {
        exercise_id: u32,
    },
    GetExerciseUpdates {
        course_id: u32,
        exercise: Vec<String>,
    },
    GetOrganization {
        organization: String,
    },
    GetOrganizations,
    GetUnreadReviews {
        course_id: u32,
    },
    LoggedIn,
    Login {
        base64: bool,
        email: Option<String>,
        set_access_token: Option<String>,
        stdin: bool,
    },
    Logout,
    MarkReviewAsRead {
        course_id: u32,
        review_id: u32,
    },
    Paste {
        exercise_id: u32,
        locale: Option<Locale>,
        paste_message: Option<String>,
        submission_path: PathBuf,
    },
    RequestCodeReview {
        exercise_id: u32,
        locale: Locale,
        message_for_reviewer: Option<String>,
        submission_path: PathBuf,
    },
    ResetExercise {
        save_old_state: bool,
        exercise_id: u32,
        exercise_path: PathBuf,
    },
    SendFeedback {
        submission_id: Option<u32>,
        feedback_url: Option<String>,
        feedback: Vec<String>,
    },
    Submit {
        dont_block: bool,
        locale: Option<Locale>,
        submission_path: PathBuf,
        exercise_id: u32,
    },
    UpdateExercises,
    WaitForSubmission {
        submission_id: u32,
    },

}

Variants

CheckExerciseUpdates

Checks for updates to any exercises that exist locally.

DownloadModelSolution

Downloads an exercise’s model solution

Fields

exercise_id: u32

The ID of the exercise.

target: PathBuf

Path to where the model solution will be downloaded.

DownloadOldSubmission

Downloads an old submission. Resets the exercise at output-path if any, downloading the exercise base from the server. The old submission is then downloaded and extracted on top of the base, using the student file policy to retain student files

Fields

submission_id: u32

The ID of the submission.

save_old_state: bool

If set, the exercise is submitted to the server before resetting it.

exercise_id: u32

The ID of the exercise.

output_path: PathBuf

Path to where the submission should be downloaded.

DownloadOrUpdateCourseExercises

Downloads exercises. If downloading an exercise that has been downloaded before, the student file policy will be used to avoid overwriting student files, effectively just updating the exercise files

Fields

download_template: bool

If set, will always download the course template instead of the latest submission, even if one exists.

exercise_id: Vec<u32>

Exercise id of an exercise that should be downloaded. Multiple ids can be given.

GetCourseData

Fetches course data. Combines course details, course exercises and course settings

Fields

course_id: u32

The ID of the course.

GetCourseDetails

Fetches course details

Fields

course_id: u32

The ID of the course.

GetCourseExercises

Lists a course’s exercises

Fields

course_id: u32

The ID of the course.

GetCourseSettings

Fetches course settings

Fields

course_id: u32

The ID of the course.

GetCourses

Lists courses

Fields

organization: String

Organization slug (e.g. mooc, hy).

GetExerciseDetails

Fetches exercise details

Fields

exercise_id: u32

The ID of the exercise.

GetExerciseSubmissions

Fetches the current user’s old submissions for an exercise

Fields

exercise_id: u32

The ID of the exercise.

GetExerciseUpdates

Checks for updates to exercises

Fields

course_id: u32

The ID of the course.

exercise: Vec<String>

An exercise. Takes two values, an exercise id and a checksum. Multiple exercises can be given.

GetOrganization

Fetches an organization

Fields

organization: String

Organization slug (e.g. mooc, hy).

GetOrganizations

Fetches a list of all organizations from the TMC server

GetUnreadReviews

Fetches unread reviews for a course

Fields

course_id: u32

The ID of the course.

LoggedIn

Checks if the CLI is authenticated. Prints the access token if so

Login

Authenticates with the TMC server and stores the OAuth2 token in config. You can log in either by email and password or an access token

Fields

base64: bool

If set, the password is expected to be a base64 encoded string. This can be useful if the password contains special characters.

email: Option<String>

The email address of your TMC account. The password will be read through stdin.

set_access_token: Option<String>

The OAUTH2 access token that should be used for authentication.

stdin: bool

If set, the password will be read from stdin instead of TTY like usual. The keyboard input is not hidden in this case, so this should only be used when running the CLI programmatically.

Logout

Logs out and removes the OAuth2 token from config

MarkReviewAsRead

Marks a review as read

Fields

course_id: u32

The ID of the course.

review_id: u32

The ID of the review.

Paste

Sends an exercise to the TMC pastebin

Fields

exercise_id: u32

The ID of the exercise.

locale: Option<Locale>

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

paste_message: Option<String>

Optional message to attach to the paste.

submission_path: PathBuf

Path to the exercise to be submitted.

RequestCodeReview

Requests code review

Fields

exercise_id: u32

The ID of the exercise.

locale: Locale

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

message_for_reviewer: Option<String>

Message for the review.

submission_path: PathBuf

Path to the directory where the submission resides.

ResetExercise

Resets an exercise. Removes the contents of the exercise directory and redownloads it from the server

Fields

save_old_state: bool

If set, the exercise is submitted to the server before resetting it.

exercise_id: u32

The ID of the exercise.

exercise_path: PathBuf

Path to the directory where the project resides.

SendFeedback

Sends feedback for an exercise submission

Fields

submission_id: Option<u32>

The ID of the submission.

feedback_url: Option<String>

The feedback answer URL.

feedback: Vec<String>

A feedback answer. Takes two values, a feedback answer id and the answer. Multiple feedback arguments can be given.

Submit

Submits an exercise. By default blocks until the submission results are returned

Fields

dont_block: bool

Set to avoid blocking.

locale: Option<Locale>

Language as a three letter ISO 639-3 code, e.g. ‘eng’ or ‘fin’.

submission_path: PathBuf

Path to the directory where the exercise resides.

exercise_id: u32

The ID of the exercise.

UpdateExercises

Updates all local exercises that have been updated on the server

WaitForSubmission

Waits for a submission to finish

Fields

submission_id: u32

The ID of the submission.