Module tmc_langs_cli::output

Contains the type definition for the output format of the CLI.

Structs

DownloadTarget

OutputData

ProjectsDirExercise

An exercise in the projects directory.

Enums

CliOutput

The format for all messages written to stdout by the CLI

DataKind

Kind

OutputResult

Status

StatusUpdateData