Crate tmc_langs_cli

CLI client for TMC

Modules

app
Create clap app

output
Output format

Structs

PrintToken

Functions

error_message_special_casing
Goes through the error chain and returns the specialized error message, if any.

into_locale

into_url

into_usize

main

print_output

run

run_checkstyle_write_results

run_core

solve_error_kind
Goes through the error chain and checks for special error types that should be indicated by the Kind.

write_result_to_file_as_json